Allow to use regex to list pages in saved selections
This allows to use a regex to select pages within a saved selection. Please
note, that this is only useful if you use manually created selections and is
not compatible with the dynmic page selection tool.

The intended use case is for example if you have a namespace where pages are
removed or added at will and the pages should be included in the PDF without
needing to select them manually one-by-one.

Signed-off-by: Frieder Schrempf <dev@fris.de>